+++ elmyra.ip.access.epo/CHANGES.rst
@@ -7,6 +7,7 @@ development
 - middleware: enhance ops usage api (differentiate between "ago" and "current" - per period)
 - middleware: uppercase patentnumber when searching at FulltextPRO
 - ui: bug: when clicking through family citations, current view state (e.g. project) is not propagated properly
+- ui: explore all family members
 
 0.90.3
 ======